requirements.txt
python-dotenv==1.0.0
requests==2.31.0
discord.py==2.5.2
twitchio==2.10.0
websockets==15.0.1
discord==2.3.2